mutex: Add (optional) `fmt::{Debug, Display}` impls
Depends on #3

I made these optional so that embedded projects that care about
optimizing for binary size can leave them disabled. They're opt-in,
rather than on by default, because I figured it would be easy for
libraries to just enable them by accident if they were default-on.

We may, also, want to consider adding a similar feature for `defmt`
support, but I'd prefer to do that separately.