mutex: prepare to release v0.2.0
It's necessary to release a breaking change version for the bump from
`mutex-traits` v0.999.0 to v1.0.0, because the crates are now no longer
compatible: v0.999.0 is not the same as v1.0, so the traits are no
longer the same. I don't love having to do this, but...since
`mutex-traits` is stable now, we'll only have to do this again if
something goes REALLY wrong.